prov/shm: move ofi_shm.h and util_shm.c code to prov/shm
No one else uses the shm specific code which is specific to the provider
and should be in the shm provider folder.
