+++ prov/shm/Makefile.include
@@ -16,7 +16,10 @@ _shm_files = \
 	prov/shm/src/smr_signal.h	\
 	prov/shm/src/smr.h		\
 	prov/shm/src/smr_dsa.h		\
-	prov/shm/src/smr_dsa.c
+	prov/shm/src/smr_dsa.c		\
+	prov/shm/src/smr_util.h		\
+	prov/shm/src/smr_util.c
+
 
 
 if HAVE_SHM_DL
