+++ prov/shm/src/smr_util.h
